+++ graphql-maven-plugin-logic/src/test/java/graphql/mavenplugin_notscannedbyspring/BasicMultiple_Client_SpringConfiguration.java
@@ -11,7 +11,9 @@ import org.springframework.context.annotation.FilterType;
 import com.graphql_java_generator.plugin.PluginMode;
 
 @Configuration
-@ComponentScan(basePackages = "com.graphql_java_generator", excludeFilters = @Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateRelaySchema.*"))
+@ComponentScan(basePackages = "com.graphql_java_generator", excludeFilters = {
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateRelaySchema.*"),
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.Merge.*") })
 public class BasicMultiple_Client_SpringConfiguration extends AbstractSpringConfiguration {
 
 	public BasicMultiple_Client_SpringConfiguration() {