+++ graphql-maven-plugin-logic/src/test/java/graphql/mavenplugin_notscannedbyspring/Basic_Client_SpringConfiguration.java
@@ -16,7 +16,9 @@ import com.graphql_java_generator.plugin.PluginMode;
  * @author etienne-sf
  */
 @Configuration
-@ComponentScan(basePackages = "com.graphql_java_generator", excludeFilters = @Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateRelaySchema.*"))
+@ComponentScan(basePackages = "com.graphql_java_generator", excludeFilters = {
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateRelaySchema.*"),
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.Merge.*") })
 public class Basic_Client_SpringConfiguration extends AbstractSpringConfiguration {
 
 	public Basic_Client_SpringConfiguration() {