+++ graphql-maven-plugin-logic/src/test/java/graphql/mavenplugin_notscannedbyspring/Github_Client_SpringConfiguration.java
@@ -7,13 +7,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.ComponentScan.Filter;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.FilterType;
 
 import com.graphql_java_generator.plugin.CustomScalarDefinition;
 import com.graphql_java_generator.plugin.PluginMode;
 
 @Configuration
-@ComponentScan(basePackages = "com.graphql_java_generator")
+@ComponentScan(basePackages = "com.graphql_java_generator", excludeFilters = {
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateRelaySchema.*"),
+		@Filter(type = FilterType.REGEX, pattern = ".*\\.Merge.*") })
 public class Github_Client_SpringConfiguration extends AbstractSpringConfiguration {
 
 	static List<CustomScalarDefinition> customScalars;