+++ graphql-maven-plugin/src/main/java/com/graphql_java_generator/mavenplugin/GraphQLMojo.java
@@ -29,6 +29,21 @@ import graphql.ThreadSafe;
 import graphql.schema.GraphQLScalarType;
 
 /**
+ * The <I>graphql</I> goal generates the java code from one or more GraphQL schemas. It allows to work in Java with
+ * graphQL, in a schema first approach.<BR/>
+ * It has two main modes:
+ * <UL>
+ * <LI><B>client mode:</B> it generates a class for each query, mutation and subscription type. These classes contain
+ * the methods to call the queries, mutations and subscriptions. That is: to execute a query against the GraphQL server,
+ * you just have to call one of this method. It also generates the POJOs from the GraphQL schema. The <B>GraphQL
+ * response is stored in these POJOs</B>, for an easy and standard use in Java.</LI>
+ * <LI><B>server mode:</B> it generates the whole heart of the GraphQL server. The developer has only to develop request
+ * to the data. That is the main method (in a jar project) or the main servler (in a war project), and all the Spring
+ * wiring, based on graphql-java-spring, itself being build on top of graphql-java. It also generates the POJOs. An
+ * option allows to annotate them with the standard JPA annotations, to make it easy to link with a database. This goal
+ * generates the interfaces for the DataFetchersDelegate (often named providers) that the server needs to implement</LI>
+ * </UL>
+ * 
  * @author etienne-sf
  */
 @Mojo(name = "graphql", defaultPhase = LifecyclePhase.GENERATE_SOURCES, requiresProject = true)